% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_util.R
\name{extend_chr}
\alias{extend_chr}
\alias{shorten_chr}
\title{Functions to handle chromosome names}
\usage{
extend_chr(chr, pfx = "chr")

shorten_chr(chr, pfx = "chr")
}
\arguments{
\item{chr}{character vector. chromosome names.}

\item{pfx}{character. prefix of a chromosome name.}
}
\value{
character vector of (extended or shortened) chromosome names.
}
\description{
\code{extend_chr} function adds a prefeix to a chromosome name if the prefix does not exist.  
\code{shorten_chr} function removes a prefix, if exists, from a chromosome names.
}
\examples{
chromosomes = c("1", "chr1", "5", "chr4", "chrX", "Y")
extend_chr(chromosomes)
shorten_chr(chromosomes)

}
