% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression_pathway_enrichment.R
\name{coexpression_pathway_enrichment}
\alias{coexpression_pathway_enrichment}
\title{Enrichment of pathways (gene sets) in a co-expression network.}
\usage{
coexpression_pathway_enrichment(
  net,
  gene_sets,
  min.gene = 3,
  max.gene = 100,
  iter = 10000
)
}
\arguments{
\item{net}{matrix. A gene x gene square matrix where each entry represents the weight of the edge between corresponding genes.}

\item{gene_sets}{list. List of pathways where each entry contains the genes in each pathway. \code{names(gene_sets)} represents the pathway name and each name must be unique.}

\item{min.gene}{integer. Minimum number of genes from a pathway available in the network.}

\item{max.gene}{integer. Maximum number of genes from a pathway available in the network.}

\item{iter}{integer. Number of random gene sets to compute the null distribution.}
}
\value{
Returns a \code{data.frame} including with the following columns.
\item{pathway}{Pathway name.}
\item{n.gene}{Number of genes available in the network.}
\item{p}{p-value for the pathway.}
\item{p.empirical}{Empirical p-value for the pathway.}
\item{fdr}{False discovery rate computed using Benjamini-Hochberg method.}
\item{fdr.empirical}{Empirical false discovery rate computed using Benjamini-Hochberg method.}
}
\description{
\code{coexpression_pathway_enrichment} computes enrichment p-values of pathways in a co-expression network.
}
\details{
To compute the enrichment p-value of a pathway in a co-expression network,
we define the score of a pathway as the the sum of weights in the co-expression network
of all possible edges between the genes in the pathway. The enrichment p-value is 
then defined as the probability that the score of the pathway is at least as big as
a random gene set with same number of genes. 

To get the null distribution, we select a number of (\code{iter}) random gene sets where 
each gene set consists of the same number of randomly selected genes, compute
their scores, and fit a normal distribution.

Enrichment of a pathway is computed only if at least \code{min.gene} and 
at most \code{max.gene} genes from the pathway are available in the network.
This cirteria helps to avoid too small or to large pathways.

Note: higher values in \code{net} should correspond to higher confidence in the edge. 
If the sign of the values mean positive or negative association between genes,
you probably should provide absolute values.
}
\examples{
genes = c('TP53', 'RBM3', 'SF3', 'LIM12', 'ATM', 'TMEM160', 'BCL2L1', 'MDM2', 'PDR', 'MEG3', 'EGFR', 'CD96', 'KEAP1', 'SRSF1', 'TSEN2')
dummy_net = matrix(rnorm(length(genes)^2), nrow = length(genes), dimnames = list(genes, genes))
dummy_net = abs((dummy_net + t(dummy_net))/2) # symmetric undirected nework
dummy_gene_sets = list(pathway1=c('TP53', 'RBM3', 'SF3', 'SF5'),
                       pathway2=c('LIM12', 'MDM2', 'BCL2L1', 'TMEM160'),
                       pathway3=c('EGFR', 'TP53', 'CD96', 'SRSF1', 'RBM14'))
enrich_res = coexpression_pathway_enrichment(net = dummy_net, gene_sets = dummy_gene_sets)
head(enrich_res)
n_sig = sum(enrich_res$fdr <= 0.05)
print(sprintf('Number of significanly enriched pathways: \%d', n_sig))
}
