% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression_shared_pathway_auc.R
\name{coexpression_shared_pathway_auc}
\alias{coexpression_shared_pathway_auc}
\title{AUC of a co-expression networks using known pathways (gene sets).}
\usage{
coexpression_shared_pathway_auc(net, gene_sets, gene_annot = NULL,
  symbol_col = "gene_name", directed = FALSE)
}
\arguments{
\item{net}{matrix. A gene x gene matrix where each entry represents the weight of the edge between corresponding genes. For a directed network, rows correspond to source genes and columns correspond to target genes. 
Row and column names of \code{net} should be HGNC gene symbols or gene ids. If they are not HGNC gene symbols, gene annotations (\code{gene_annot}) must be provided.}

\item{gene_sets}{list. List of pathways where each entry contains the genes (HGNC gene symbols) in each pathway.}

\item{gene_annot}{data.frame. Gene annotations, where the row names correspond to gene ids and one of it columns (\code{symbol_col}) contains HGNC gene symbols.}

\item{symbol_col}{character or integer. Name or index of the column containg HGNC gene symbol in \code{gene_annot}.}

\item{directed}{logical. Directed interactions? If TRUE, it is assumed that the g1_col gene regulates the g2_col gene.}
}
\value{
Returns a list object with the following items.
\item{pr}{Precision curve object.}
\item{roc}{Roc curve object.}
\item{adjacency}{An adjacency matrix representing if a pair of genes have at least one common pathway.}
}
\description{
This function computes area under the curve (AUC) of a co-expression network using known pathways (gene sets).
}
\details{
This function computes the area under the ROC curve and the precision-recall curve. 
The area is computed using the \code{PRROC} package considering an interaction between 
a pair of genes known if the genes have at least one pathway in common. 
Note: higher values in \code{net} should correspond to higher confidence in the edge. 
If the sign of values mean positive or negative association between genes, 
you probably should provide absolute values. 

This function does not yet handle directed networks.
}
\examples{
genes = c('TP53', 'RBM3', 'SF3', 'LIM12', 'ATM', 'TMEM160', 'BCL2L1', 'MDM2', 'PDR', 'MEG3', 'EGFR', 'CD96', 'KEAP1', 'SRSF1', 'TSEN2')
dummy_net = matrix(rnorm(length(genes)^2), nrow = length(genes), dimnames = list(genes, genes))
dummy_net = abs((dummy_net + t(dummy_net))/2) # symmetric undirected nework
gsea_fn = "your_gene_sets_file.gmt" # examples available at http://software.broadinstitute.org/gsea/msigdb/collections.jsp (download gene symbols GMT files)
gene_sets = read_gsea_db(gsea_fn)
auc_res = coexpression_shared_pathway_auc(net = dummy_net, gene_sets = gene_sets)
print(sprintf('AUC under ROC curve: \%s', auc_res$roc$auc))
print(sprintf('AUC under PR curve: \%s', auc_res$pr$auc.integral))
plot(auc_res$roc)
plot(auc_res$pr)

}
