% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_expression_data.R
\name{filter_expr_on_chr}
\alias{filter_expr_on_chr}
\alias{filter_expr_on_mappability}
\alias{filter_expr_on_gene_type}
\alias{filter_expr_on_tpm_read}
\alias{filter_expr_on_variance}
\alias{filter_expr_on_coeff_of_variation}
\title{Functions to filter gene expression data}
\usage{
filter_expr_on_chr(expr.df, annot.gene, chr.include = as.character(1:22),
  chr.exclude = NULL, chr.col = "chr")

filter_expr_on_mappability(expr.df, annot.mappability,
  min.mappability = 0.97, mappability.col = "mappability")

filter_expr_on_gene_type(expr.df, annot.gene, type.col = "gene_type",
  type.values = "protein_coding")

filter_expr_on_tpm_read(expr.df, tpm.df, count.df, min.tpm = 0.1,
  min.count = 6, min.samples = 10)

filter_expr_on_variance(expr.df, raw.df, n, min.var = 1e-06,
  min.mean = -Inf)

filter_expr_on_coeff_of_variation(expr.df, raw.df, n, min.var = 1e-06,
  min.mean = 0.01)
}
\arguments{
\item{expr.df}{data.frame. processed expression data (gene x sample). See details.}

\item{annot.gene}{data.frame. gene annotation data frame with columns 'gene_id', 'gene_name', 'chr', 'start_pos', 'end_pos', etc.  See details.}

\item{chr.include}{character vector. chromosomoes to include.}

\item{chr.exclude}{character vector. chromosomoes to exclude.}

\item{chr.col}{character or numeric. column name or column number containing chromosomes of genes in the gene annotation data frame.}

\item{annot.mappability}{data.frame. gene mappability data frame with at columns 'gene_id', 'mappability', etc. See details.}

\item{min.mappability}{numeric. minimum gene mappability.}

\item{mappability.col}{character or numeric. column name or column number containing mappability of genes.}

\item{type.col}{character or numeric. column name or column number containing gene types in the gene annotation data frame.}

\item{type.values}{character vector. target gene types (e.g., protein_coding) to include.}

\item{tpm.df}{data.frame. raw (unprocessed) TPM data (gene x sample). See details.}

\item{count.df}{data.frame. raw (uncrocessed) read-count data (gene x sample). See details.}

\item{min.tpm}{numeric. minimum TPM.}

\item{min.count}{numeric. minimum read count.}

\item{min.samples}{numeric. minimum number of samples.}

\item{raw.df}{data.frame. any type of raw (uncrocessed) data (gene x sample). See details.}

\item{n}{maximum number of genes.}

\item{min.var}{numeric. minimum variance of expression per gene across all samples.}

\item{min.mean}{numeric. minimum mean of expression per gene across all samples.}
}
\value{
Each function returns a data frame with filtered expression data.
}
\description{
Each function filters gene expression data (gene x sample) based on a defined criteria.
}
\details{
Above functions filter processed gene expression data (\code{expr.df}) based on several different criteria.
Typical filtering scenarios include:
\itemize{
  \item keeping only autosomal genes, excluding genes in sex and mitochondrial chromosomes using \code{filter_expr_on_chr},
  \item including genes with mappability >= 0.8 for an analysis using \code{filter_expr_on_mappability},
  \item inlcluding only protein-coding genes in an analysis using \code{filter_expr_on_gene_type},
  \item include genes with TPM > 0.1 in \eqn{\ge} 10 samples and read count > 6 in \eqn{\ge} 10 samples using \code{filter_expr_on_tpm_read},
  \item take only 500 most variant genes using \code{filter_expr_on_variance}, and
  \item take only 500 genes with highest coefficient of variation using \code{filter_expr_on_coeff_of_variation}.
}

Data formats are important for these filter functions.

Any expression-like data (\code{expr.df}, \code{raw.df}, \code{tpm.df}, or \code{count.df}) must be in a data.frame or a matrix-like object, 
where each row represents a gene and each column represents a sample (or an individual).
The data frame must have row-names with unique gene ids and column-names with unique sample ids.

Any gene annotation data (\code{annot.gene}, or \code{annot.mappability}) must be in a data.frame or a matrix-like object, 
where each row represent a gene, and each column represents a property of the gene.
The data frame must have row-names with unique gene ids matching to gene ids in expression-like data.
An annotation data frame may contain more than necessary columns,
and necessary columns can be configured in each function.
}
\section{Functions}{
\itemize{
\item \code{filter_expr_on_chr}: filteres to include or explore genes from given chromosomes (if not NULL). Chromosome names with or without "chr" prefix are gracefully handled.

\item \code{filter_expr_on_mappability}: filters out genes with mappability less than a threshold.

\item \code{filter_expr_on_gene_type}: filters out non protein-coding genes from expression data.

\item \code{filter_expr_on_tpm_read}: filters expression data to include each gene with TPM >= \code{min.tpm} in >= \code{min.samples} samples and with read count >= \code{min.count} in >= \code{min.samples} samples.

\item \code{filter_expr_on_variance}: filters expression data to include most variant \code{n} genes with mean >= \code{min.mean} and variance >= \code{min.var}

\item \code{filter_expr_on_coeff_of_variation}: filters expression data to include genes with \code{n} highest coefficient of variation (variance/mean) and mean >= \code{min.mean} and variance >= \code{min.var}
}}

