% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_string_ppi.R
\name{get_string_ppi}
\alias{get_string_ppi}
\title{Protein-protein interactions (PPIs) within given set of genes.}
\usage{
get_string_ppi(genes, directed = FALSE, string_version = "10",
  species = 9606, score_threshold = 400, string_dir = "",
  max_homology_bitscore = Inf, benchmark_pathway = NULL)
}
\arguments{
\item{genes}{A character vector with gene symbols.}

\item{directed}{logical. Directed interactions?}

\item{string_version}{character. Version of STRING database.}

\item{species}{numeric. Species code for STRING (9606 for homo sapiens).}

\item{score_threshold}{numeric. Interactions in STRING with score >= score_threshold are considered true interactions.}

\item{string_dir}{character. Local path where STRING database will be stored locally so that it can be used offline. If \code{string_dir = ""}, a temporary directory is used.}

\item{max_homology_bitscore}{numeric. Maximum homology score. \code{0} to filter all interactions between homologous proteins, \code{Inf} not to filter any interaction.}

\item{benchmark_pathway}{character. Pathway type to benchmark. \code{NULL} to use all pathway types. Examples: \code{"KEGG", "REACTOME", "BIOCARTA", "Disease", "Pfam", "NCI", "ECOCYC".} See \code{pathwayType} parameter in \code{STRINGdb::benchmark_ppi()} function for details.}
}
\value{
Returns a data frame with interactions between the given genes.
}
\description{
This function returns PPIs available in STRING database within a given set of genes.
}
\details{
Note: all the genes are expected as HGNC gene symbols.
}
\examples{
genes = c("TP53", "RBM3", "SF3", "LIM12", "MDM4", "TMEM160", "TP53BP2", "MDM2", "PDR", "MEG3", "EGFR")
interactions = get_string_ppi(genes)
}
