% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_genes_to_string_id.R
\name{map_genes_to_string_ids}
\alias{map_genes_to_string_ids}
\title{Map genes to STRING protein IDs.}
\usage{
map_genes_to_string_ids(
  genes,
  string_version = "10",
  species = 9606,
  string_dir = ""
)
}
\arguments{
\item{genes}{A character vector with HGNC gene symbols.}

\item{string_version}{character. Version of STRING database.}

\item{species}{numeric. Species code for STRING (9606 for homo sapiens).}

\item{string_dir}{character. Local path where STRING database will be stored locally so that it can be used offline. If \code{string_dir = ""}, a temporary directory is used.}
}
\value{
a character vector with mapped STRING ids.
}
\description{
This function maps HGNC gene symbols to protein ids in STRING.
}
\examples{
genes = c('TP53', 'RBM3', 'SF3', 'LIM12', 'ATM', 'TMEM160', 'BCL2L1')
string_ids = map_genes_to_string_ids(genes)
}
