% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gsea_db.R
\name{read_gsea_db}
\alias{read_gsea_db}
\title{Function to read GSEA gene sets}
\usage{
read_gsea_db(fn)
}
\arguments{
\item{fn}{either a character string or a connection object to read.}
}
\value{
Returns a list of gene sets. Each item in the list is a character vector representing a gene set. The name of each gene set is available as the name of each item in the list. 
Note: the 2nd column in the file containing an url of the gene set is ignored.
}
\description{
This function reads a file with gene sets in gmt format. Representive examples are found \href{http://software.broadinstitute.org/gsea/msigdb/collections.jsp}{here}.
}
