% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_ppi_enrichment.R
\name{string_ppi_enrichment}
\alias{string_ppi_enrichment}
\title{Enrichment of protein-protein interactions (PPIs)}
\usage{
string_ppi_enrichment(interaction_df, bg, directed = FALSE,
  g1_col = colnames(interaction_df)[1],
  g2_col = colnames(interaction_df)[2], n_iter = 1000,
  string_version = "10", species = 9606, score_threshold = 400,
  string_dir = "", max_homology_bitscore = Inf,
  benchmark_pathway = NULL)
}
\arguments{
\item{interaction_df}{A data.frame with gene-gene interactions. It must have at least 2 columns containing gene symbols.}

\item{bg}{A character vector with background gene symbols. It should contain all genes in the interaction data frame and other genes analyzed to discover the interactions.}

\item{directed}{logical. Directed interactions? If TRUE, it is assumed that the g1_col gene regulates the g2_col gene.}

\item{g1_col}{character. Column name of interaction_df that contains one gene of an interaction.}

\item{g2_col}{character. Column name of interaction_df that contains another gene of an interaction.}

\item{n_iter}{numeric. Number of random iterations to compute the background distribution.}

\item{string_version}{character. Version of STRING database.}

\item{species}{numeric. Species code for STRING (9606 for homo sapiens).}

\item{score_threshold}{numeric. Interactions in STRING with score >= score_threshold are considered true interactions.}

\item{string_dir}{character. Local path where STRING database will be stored locally so that it can be used offline.}

\item{max_homology_bitscore}{numeric. Maximum homology score. \code{0} to filter all interactions between homologous proteins, \code{Inf} not to filter any interaction.}

\item{benchmark_pathway}{character. Pathway type to benchmark. \code{NULL} to use all pathway types. Examples: \code{"KEGG", "REACTOME", "BIOCARTA", "Disease", "Pfam", "NCI", "ECOCYC".} See \code{pathwayType} parameter in \code{STRINGdb::benchmark_ppi()} function for details.}
}
\value{
Returns a list object with the following items.
\item{pvalue}{numeric, enrichment p-value}
\item{random_known_ppi}{numeric vector, empirical number of known PPIs.}
\item{observed_known_ppi}{numeric, observed number of known PPIs in the given interaction set.}
\item{total_known_ppi}{numeric, total number of known PPIs possible between background genes.}
}
\description{
This function computes enrichment of know PPIs in a given set of interactions.
}
\details{
This function computes enrichment of known protein-protein interactions (PPIs) in the given set of interactions. 
It computes enrichment p-value as the probably that the number of known PPIs in the given set 
is greater than or equal to that in a random set of same size.

Note: all the genes (both in \code{interaction_df} and \code{bg}) are expected as HGNC gene symbols.
}
\examples{
gene1s = c('TP53', 'TP53', 'RBM3', 'SF3', 'LIM12', 'MDM4', 'TMEM160')
gene2s = g2=c('TP53BP2', 'MDM2', 'PDR', 'MEG3', 'TP53', 'MDM2', 'EGFR')
interaction_df = data.frame(g1 = gene1s, g2 = gene2s, stringsAsFactors = F)
bg = unique(c(gene1s, gene2s, c('CD96', 'KEAP1', 'SRSF1', 'TSEN2')))
enrich_res = string_ppi_enrichment(interaction_df, bg)
}
