% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmm_normalize.R
\name{tmm_normalize}
\alias{tmm_normalize}
\title{TMM-normalization of RNA-seq data}
\usage{
tmm_normalize(expr_mat)
}
\arguments{
\item{expr_mat}{matrix. Data matrix where rows and columns represent genes and samples, respectively.}
}
\value{
Returns a transformed matrix.
}
\description{
This function normalizes a gene expression matrix using TMM.
}
\examples{
y <- matrix( rpois(1000, lambda=5), nrow=200 )
tmm_y = tmm_normalize(y)
}
\references{
\insertRef{Robinson2010}{genomicsutil}
}
