% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_inv_normal.R
\name{to_inv_normal}
\alias{to_inv_normal}
\title{Transformation data to inverse normal distribution}
\usage{
to_inv_normal(expr_mat)
}
\arguments{
\item{expr_mat}{matrix. Data matrix.}
}
\value{
Returns a transformed matrix.
}
\description{
This function transforms values in each row of a matrix to a rank based inverse normal transformation.
}
\examples{
x = matrix(1:100, nrow=5, byrow=T, dimnames = list(paste0('gene', 1:5), paste0('sample', 1:20)))
inv_norm_x = to_inv_normal(x)
}
